Split SvgLineSegment
