+++ Source/Paths/SvgLineSegment.Drawing.cs
@@ -1,0 +1,24 @@
+#if !NO_SDC
+using System.Drawing;
+using System.Drawing.Drawing2D;
+
+namespace Svg.Pathing
+{
+    public sealed partial class SvgLineSegment : SvgPathSegment
+    {
+        public override PointF AddToPath(GraphicsPath graphicsPath, PointF start, SvgPathSegmentList parent)
+        {
+            var end = ToAbsolute(End, IsRelative, start);
+            graphicsPath.AddLine(start, end);
+            return end;
+        }
+
+        [System.Obsolete("Use new AddToPath.")]
+        public override void AddToPath(GraphicsPath graphicsPath)
+        {
+            AddToPath(graphicsPath, Start, null);
+        }
+
+    }
+}
+#endif
