+++ Source/Paths/SvgLineSegment.cs
@@ -1,31 +1,13 @@
 using System.Drawing;
-#if !NO_SDC
-using System.Drawing.Drawing2D;
-#endif
 
 namespace Svg.Pathing
 {
-    public sealed class SvgLineSegment : SvgPathSegment
+    public sealed partial class SvgLineSegment : SvgPathSegment
     {
         public SvgLineSegment(bool isRelative, PointF end)
             : base(isRelative, end)
         {
         }
-
-#if !NO_SDC
-        public override PointF AddToPath(GraphicsPath graphicsPath, PointF start, SvgPathSegmentList parent)
-        {
-            var end = ToAbsolute(End, IsRelative, start);
-            graphicsPath.AddLine(start, end);
-            return end;
-        }
-
-        [System.Obsolete("Use new AddToPath.")]
-        public override void AddToPath(GraphicsPath graphicsPath)
-        {
-            AddToPath(graphicsPath, Start, null);
-        }
-#endif
 
         public override string ToString()
         {
